Refactoring, controllers, request, emial, olicies, add traitsand checking similar jobs
