+++ app/Http/Requests/Api/V1/Auth/RegisterUserRequest/StoreUserRequest.php
@@ -1,0 +1,54 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Http\Requests\Api\V1\Auth\RegisterUserRequest;
+
+use App\Models\User;
+use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Foundation\Http\FormRequest;
+
+class StoreUserRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'first_name' => ['required', 'string', 'max:255'],
+            'last_name' => ['required', 'string', 'max:255'],
+            'user_name' => ['required', 'string', 'max:255', 'alpha_dash', 'unique:users,user_name'],
+            'email' => ['required', 'email', 'unique:users,email'],
+            'password' => ['required', 'string', 'min:6'],
+            'password_confirmation' => ['required', 'string', 'min:6'],
+        ];
+    }
+
+    /**
+     * Store the user
+     */
+    public function storeUser(): User
+    {
+        return User::create([
+            'first_name' => $this->first_name,
+            'last_name' => $this->last_name,
+            'user_name' => $this->user_name,
+            'slug' => Str::slug($this->user_name),
+            'email' => $this->email,
+            'password' => Hash::make($this->password),
+            'remember_token' => Str::random(20),
+        ]);
+    }
+}
