+++ app/Http/Requests/Api/V1/TokenRequest/TokenRequest.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Requests\Api\V1\TokenRequest;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class TokenRequest extends FormRequest
+{
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'email' => 'required|email',
+            'password' => 'required'
+        ];
+    }
+}
