--- app/Http/Requests/JobPostsRequest/UpdateJobPostRequest.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace App\Http\Requests\JobPostsRequest;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class UpdateJobPostRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return false;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
-    public function rules(): array
-    {
-        return [
-            //
-        ];
-    }
-}
