--- app/Http/Requests/LoginUserRequest/LoginUserRequest.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace App\Http\Requests\LoginUserRequest;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class LoginUserRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
-    public function rules(): array
-    {
-        return [
-            'email' => ['required', 'email', 'exists:users'],
-            'password' => ['required', 'min:6'],
-        ];
-    }
-}
