--- app/Http/Requests/TokenRequest/TokenRequest.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace App\Http\Requests\TokenRequest;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class TokenRequest extends FormRequest
-{
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
-    public function rules(): array
-    {
-        return [
-            'email' => 'required|email',
-            'password' => 'required'
-        ];
-    }
-}
