+++ app/Http/Resources/Api/V1/UserResource/UserResource.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Http\Resources\Api\V1\UserResource;
+
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class UserResource extends JsonResource
+{
+    // public static $wrap = 'user';
+    /**
+     * Transform the resource into an array.
+     *
+     * @return array<string, mixed>
+     */
+    public function toArray(Request $request): array
+    {
+        return [
+            'type' => 'users',
+            'id' => $this->id,
+            'attributes' => [
+                'first_name' => $this->first_name,
+                'last_name' => $this->last_name,
+                'user_name' => $this->user_name,
+                'slug' => Str::slug($this->user_name),
+                'email' => $this->email,
+                'email_verified_at' => $this->email_verified_at,
+                'role' => $this->role,
+                'status' => $this->status,
+                'created_at' => $this->created_at,
+                'updated_at' => $this->updated_at,
+            ],
+        ];
+    }
+}
