--- app/Http/Resources/JobPostResource.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace App\Http\Resources;
-
-use Illuminate\Support\Str;
-use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\JsonResource;
-
-class JobPostResource extends JsonResource
-{
-    /**
-     * Transform the resource into an array.
-     *
-     * @return array<string, mixed>
-     */
-    public function toArray(Request $request): array
-    {
-        return [
-            'id' => $this->id,
-            'user_id' => $this->user_id,
-            'title' => $this->title,
-            'slug' => Str::slug($this->title),
-            'description' => $this->description,
-            'requirements' => $this->requirements,
-            'location' => $this->location,
-            'department' => $this->department,
-            'employment_type' => $this->employment_type,
-            'status' => $this->status,
-            'created_at' => $this->created_at,
-            'updated_at' => $this->updated_at,
-        ];
-    }
-}
