+++ app/Providers/Api/V1/AppServiceProvider/AppServiceProvider.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Providers\Api\V1\AppServiceProvider;
+
+use Illuminate\Support\ServiceProvider;
+
+class AppServiceProvider extends ServiceProvider
+{
+    /**
+     * Register any application services.
+     */
+    public function register(): void
+    {
+        //
+    }
+
+    /**
+     * Bootstrap any application services.
+     */
+    public function boot(): void
+    {
+        //
+    }
+}
