+++ database/seeders/JobPostSeeder.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Database\Seeders;
 
-use App\Models\User;
 use App\Models\JobPost;
 use Illuminate\Database\Seeder;
